requirements.txt
aiogram~=3.0
fastapi~=0.104
uvicorn[standard]~=0.24
python-dotenv~=1.0